Add information for composer version 1 and PHP8
When the current version of composer (v2) is used, installing the horde dev-depencies fails with error message:
"The PEAR repository has been removed from Composer 2.x"
This commit adds information to use Composer version 1.x.

Also PHP 8 is currently not possible, as installing the dev-depencies fails with the follwoing error message for
a lot of horde dev-requirements:
"requires php <8.0.0.0 -> your PHP version (8.0.14) does not satisfy that requirement."
This commit adds the information to use the flag `--ignore-platform-reqs`.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>